+++ KVMultiDet/analysis/LinkDef.h
@@ -14,8 +14,7 @@
 #pragma link C++ class KVBatchJob;
 #pragma link C++ class KVGEBatchJob;
 #pragma link C++ class KV_CCIN2P3_GE;
-#pragma link C++ class KVRootBatch;
-#pragma link C++ class KVLinuxBatch;
+#pragma link C++ class KVXtermBatch;
 #pragma link C++ class KVPROOFLiteBatch;
 #pragma link C++ class KVDataAnalyser+;
 #pragma link C++ class KVDataSetAnalyser+;